Put serilize/deserialize in same class
